+++ astroquery/mast/tests/test_mast_remote.py
@@ -783,6 +783,45 @@ class TestMast:
         with caplog.at_level("INFO", logger="astroquery"):
             assert "products were duplicates" in caplog.text
 
+    def test_observations_download_products_cloud(self, tmp_path, msa_product_table):
+        pytest.importorskip('boto3')
+
+        Observations.enable_cloud_dataset()
+
+        # Adding a product that's not in the cloud to test mixed downloads
+        in_uri = "mast:IUE/url/pub/vospectra/iue2/swp18830mxlo_vo.fits"
+        new_row = {col: msa_product_table[col][0] for col in msa_product_table.colnames}
+        new_row["dataURI"] = in_uri
+        new_row["productFilename"] = Path(in_uri).name
+        msa_product_table = msa_product_table.copy()
+        msa_product_table.add_row(new_row)
+
+        with pytest.warns(NoResultsWarning, match='Skipping download'):
+            result = Observations.download_products(msa_product_table,
+                                                    download_dir=tmp_path,
+                                                    cloud_only=True)
+        assert isinstance(result, Table)
+        assert len(result) == 2
+        # First product should be downloaded from cloud, second should be skipped
+        assert result['Status'][0] == 'COMPLETE'
+        assert result['Status'][1] == 'SKIPPED'
+        assert Path(result['Local Path'][0]).exists()
+        assert not Path(result['Local Path'][1]).exists()
+        Path.unlink(result['Local Path'][0])  # clean up file
+
+        # Should fall back and download if cloud_only is False
+        with pytest.warns(InputWarning, match='Falling back to MAST download.'):
+            result = Observations.download_products(msa_product_table,
+                                                    download_dir=tmp_path)
+        assert isinstance(result, Table)
+        assert len(result) == 2
+        assert result['Status'][0] == 'COMPLETE'
+        assert result['Status'][1] == 'COMPLETE'
+        assert Path(result['Local Path'][0]).exists()
+        assert Path(result['Local Path'][1]).exists()
+
+        Observations.disable_cloud_dataset()
+
     def test_observations_download_file(self, tmp_path):
 
         def check_result(result, path):
@@ -820,7 +859,8 @@ class TestMast:
         check_result(result, local_path_file)
 
     @pytest.mark.parametrize("in_uri", [
-        'mast:HLA/url/cgi-bin/getdata.cgi?download=1&filename=hst_05206_01_wfpc2_f375n_wf_daophot_trm.cat',
+        'mast:GALEX/url/data/GR6/pipe/01-vsn/03329-MISDR1_18916_0459/d/01-main/0001-img/07-try/'
+        'MISDR1_18916_0459-fd-flagstar.fits.gz',
         'mast:HST/product/u24r0102t_c3m.fits'
     ])
     def test_observations_download_file_cloud(self, tmp_path, in_uri):
@@ -832,6 +872,29 @@ class TestMast:
         result = Observations.download_file(uri=in_uri, cloud_only=True, local_path=tmp_path)
         assert result == ('COMPLETE', None, None)
         assert Path(tmp_path, filename).exists()
+
+        Observations.disable_cloud_dataset()
+
+    def test_observations_download_file_cloud_not_found(self, tmp_path):
+        pytest.importorskip("boto3")
+        in_uri = 'mast:IUE/url/pub/vospectra/iue2/swp18830mxlo_vo.fits'
+
+        Observations.enable_cloud_dataset()
+
+        # Warn and fallback
+        with pytest.warns(InputWarning, match='Falling back to MAST download.'):
+            result = Observations.download_file(uri=in_uri, local_path=tmp_path)
+            assert result == ('COMPLETE', None, None)
+            assert Path(tmp_path, Path(in_uri).name).exists()
+            Path.unlink(Path(tmp_path, Path(in_uri).name))  # clean up file
+
+        # Skip if cloud_only is set
+        with pytest.warns(NoResultsWarning, match='Skipping download'):
+            result = Observations.download_file(uri=in_uri, cloud_only=True, local_path=tmp_path)
+            assert result == ('SKIPPED', None, None)
+            assert not Path(tmp_path, Path(in_uri).name).exists()
+
+        Observations.disable_cloud_dataset()
 
     def test_observations_download_file_escaped(self, tmp_path):
         # test that `download_file` correctly escapes a URI
@@ -863,6 +926,16 @@ class TestMast:
         assert result == ("COMPLETE", None, None)
         assert Path(tmp_path, filename).exists()
 
+    def test_observations_list_cloud_missions(self):
+        pytest.importorskip('boto3')
+        Observations.enable_cloud_dataset()
+        missions = Observations.list_cloud_datasets()
+        assert isinstance(missions, list)
+        assert len(missions) > 0
+        for m in ['hst', 'jwst', 'panstarrs', 'galex', 'tess']:
+            assert m in missions
+        Observations.disable_cloud_dataset()
+
     @pytest.mark.parametrize("test_data_uri, expected_cloud_uri", [
         ("mast:HST/product/u24r0102t_c1f.fits",
          "s3://stpubdata/hst/public/u24r/u24r0102t/u24r0102t_c1f.fits"),
@@ -888,6 +961,8 @@ class TestMast:
         uri = Observations.get_cloud_uri(test_data_uri)
         assert uri == expected_cloud_uri, f'Cloud URI does not match expected. ({uri} != {expected_cloud_uri})'
 
+        Observations.disable_cloud_dataset()
+
     @pytest.mark.parametrize("test_obs_id", ["25568122", "31411", "107604081"])
     def test_observations_get_cloud_uris(self, test_obs_id):
         pytest.importorskip("boto3")
@@ -911,6 +986,8 @@ class TestMast:
         with pytest.warns(NoResultsWarning):
             Observations.get_cloud_uris(products,
                                         extension='png')
+
+        Observations.disable_cloud_dataset()
 
     def test_observations_get_cloud_uris_list_input(self):
         pytest.importorskip("boto3")
@@ -941,8 +1018,10 @@ class TestMast:
                                         extension='png')
 
         # check for warning if one of the URIs is not found
-        with pytest.warns(NoResultsWarning, match='Failed to retrieve MAST relative path'):
+        with pytest.warns(NoResultsWarning, match='Failed to retrieve cloud path'):
             Observations.get_cloud_uris(['mast:HST/product/does_not_exist.fits'])
+
+        Observations.disable_cloud_dataset()
 
     def test_observations_get_cloud_uris_query(self):
         pytest.importorskip("boto3")
@@ -969,6 +1048,8 @@ class TestMast:
         with pytest.warns(NoResultsWarning):
             Observations.get_cloud_uris(target_name=234295611)
 
+        Observations.disable_cloud_dataset()
+
     def test_observations_get_cloud_uris_no_duplicates(self, msa_product_table):
         pytest.importorskip("boto3")
 
@@ -983,6 +1064,8 @@ class TestMast:
         # Check that only one URI is returned
         uris = Observations.get_cloud_uris(products)
         assert len(uris) == 1
+
+        Observations.disable_cloud_dataset()
 
     ######################
     # CatalogClass tests #